Revert "Reuse some codes"
This reverts commit 25549273d0abd7e1bd6a980e520224b597430a53.
